refactor: rename core scripts and its usages
* rename core scripts

* rename core script references

* rename core script reference

* refactor: rename core script usage
