+++ src/noisepy/seis/main.py
@@ -20,9 +20,9 @@ from .channelcatalog import CSVChannelCatalog, XMLStationChannelCatalog
 from .constants import CONFIG_FILE, STATION_FILE
 from .datatypes import Channel, ConfigParameters
 from .numpystore import NumpyCCStore, NumpyStackStore
-from .S0A_download_ASDF_MPI import download
-from .S1_fft_cc_MPI import cross_correlate
-from .S2_stacking import stack
+from .download import download
+from .correlate import cross_correlate
+from .stack import stack
 from .scedc_s3store import SCEDCS3DataStore
 from .scheduler import (
     AWSBatchArrayScheduler,
