+++ src/noisepy/seis/noise_module.py
@@ -26,7 +26,7 @@ from scipy.fftpack import next_fast_len
 from scipy.signal import hilbert
 
 from noisepy.seis.datatypes import ChannelData, StackMethod
-from noisepy.seis.S1_fft_cc_MPI import ConfigParameters
+from noisepy.seis.correlate import ConfigParameters
 
 logger = logging.getLogger(__name__)
 """
